src/requirements-buster.txt
Adafruit-Blinka==6.18.0
adafruit-circuitpython-busdevice==5.1.1
adafruit-circuitpython-motor==3.3.4
adafruit-circuitpython-pca9685==3.3.9
adafruit-circuitpython-register==1.9.7
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PlatformDetect==3.19.1
Adafruit-PureIO==1.1.9
appdirs==1.4.4
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
cbor2==5.4.2.post1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
Deprecated==1.2.13
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
Flask-Cors==3.0.10
fourletterphat==0.1.0
gpiozero==1.6.2
gTTS==2.2.3
html5lib==1.0.1
idna==2.6
imutils==0.5.4
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
luma.core==2.3.1
luma.oled==3.8.1
lxml==4.3.2
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mpu6050-raspberrypi==1.2
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
opencv-contrib-python==4.1.0.25
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
psutil==5.5.1
pybase64==1.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.53.3
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
pyusb==1.2.1
pyxdg==0.25
pyzmq==22.3.0
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simpleaudio==1.0.4
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
smbus2==0.4.1
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
sysv-ipc==1.1.0
thonny==3.3.10
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
unicornhathd==0.0.4
urllib3==1.24.1
webencodings==0.5.1
websockets==10.3
Werkzeug==0.14.1
wrapt==1.10.11
zmq==0.0.0
